Sprout `EmbeddableForm` Furniture
As we begin this, we thought that it makes sense to start down the path
of using a generic piece of Furniture, so that we're catfooding what the
customer experience for other Spaces look like.

So we've sprouted a piece of furniture that can embed a form from
Airtable right into a Space! MAGIC!

Co-authored-by: Ana Ulin <ana@ulin.org>